refactored native code, changeed structure of src directory
