+++ ios/RnExecutorch.h
@@ -3,9 +3,7 @@
 
 #ifdef RCT_NEW_ARCH_ENABLED
 #import <RnExecutorchSpec/RnExecutorchSpec.h>
-#endif
 
-#ifdef RCT_NEW_ARCH_ENABLED
 @interface RnExecutorch : RCTEventEmitter <NativeRnExecutorchSpec>
 
 @end
