+++ package.json
@@ -187,7 +187,7 @@
   "codegenConfig": {
     "name": "RnExecutorchSpec",
     "type": "all",
-    "jsSrcsDir": "./src",
+    "jsSrcsDir": "./src/native",
     "android": {
       "javaPackageName": "com.swmansion.rnexecutorch"
     }
